公益增加查询接口

+++ src/main/java/com/foodie/portal/publicbenefit/CreatePublicBenefitCommand.java
@@ -16,5 +16,5 @@ public class CreatePublicBenefitCommand {
 
     private String title;
     private String content;
-    private BigDecimal totalFoundation;
+    private double totalFoundation;
 }
